1-Q&A-Chatbots/requirements.txt
langchain
langchain-openai
python-dotenv
langchain-community
ipykernel
langchain-core
streamlit